Update build to include all keys, toggleable UI elements, and improved cursor handle
Summary:
This includes a bunch of unreviewed changes, but it should finish off all the behavioers we need
for user testing. It includes:

- iOS animation fixes (D27148, D27149)
- Design tweaks: removing the drop shadow, corner decals, and swapping out assets (D27145)
- Dynamic iconography for the 'Dismiss'/'Jump Out' variant (D27135, D27136, D27138)
- Toggleable UI elements (D27143, D27151)
- Cursor improvements (D27124, D27142, D27153)
- The 'Toggle Sign'/'+-' functionality for the fraction keypad (D27154 - highly experimental and untested)

There are also a number of cheap-to-moderate performance optimizations that haven't even made it to
review but are being included in the build to make it useable (see: the [origin/optimize](https://github.com/khan/math-input/tree/optimize)).

Test Plan: Run through the various test plans!

Auditors: kevinb